+++ src/unnamed/FUN_0040b180.c
@@ -7,7 +7,7 @@
 #include "ghidra_types.h"
 
 
-int FUN_0040b180(undefined4 param_1,int param_2,byte param_3)
+int FUN_0040b180(undefined4 param_1,int param_2,int param_3)
 
 {
   int iVar1;
